requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper==20231117
transformers==4.40.1
torch==2.2.2
moviepy==1.0.3
pydub==0.25.1
jinja2==3.1.3
numpy==1.26.4
